Vowels
Count vowels in a string, every letter in the english alphabet is either a vowel or consonant, [a, e, i, o, u] is a list of all vowels. The rest are all considered consonants. 

A function that takes a string and returns a tuple containing a new string made of all and only the  vowels from the original  string and the number of duplicates in the original string. Only the first instance of the vowel is considered.